added hook for includeDescedants logic
