+++ assets/js/src/core/modules/asset/editor/types/folder/tab-manager/tabs/list/sidebar/filters/hooks/use-include-descendants-filter.ts
@@ -1,0 +1,41 @@
+/**
+* Pimcore
+*
+* This source file is available under two different licenses:
+* - Pimcore Open Core License (POCL)
+* - Pimcore Commercial License (PCL)
+* Full copyright and license information is available in
+* LICENSE.md which is distributed with this source code.
+*
+*  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+*  @license    https://github.com/pimcore/studio-ui-bundle/blob/1.x/LICENSE.md POCL and PCL
+*/
+
+import { useState } from 'react'
+import type { CheckboxChangeEvent } from 'antd/es/checkbox'
+import { useFilters } from './use-filters'
+
+interface IUseIncludeDescendantsFilterHookReturn {
+  isIncludeDescendants: boolean
+  setIsIncludeDescendants: (value: boolean) => void
+  handleChangeIsIncludeDescendants: (e: CheckboxChangeEvent) => void
+}
+
+export const useIncludeDescendantsFilter = (): IUseIncludeDescendantsFilterHookReturn => {
+  const [isIncludeDescendants, setIsIncludeDescendants] = useState<boolean>(false)
+
+  const { updateIsIncludeDescendants } = useFilters()
+
+  const handleChangeIsIncludeDescendants = (e: CheckboxChangeEvent): void => {
+    const includeDescendantsValue = e.target.checked
+
+    setIsIncludeDescendants(includeDescendantsValue)
+    updateIsIncludeDescendants(includeDescendantsValue)
+  }
+
+  return {
+    isIncludeDescendants,
+    setIsIncludeDescendants,
+    handleChangeIsIncludeDescendants
+  }
+}
